+++ pkg/querier/blocks_store_queryable.go
@@ -332,7 +332,7 @@ func (q *blocksStoreQuerier) Select(ctx context.Context, _ bool, sp *storage.Sel
 	return q.selectSorted(ctx, sp, tenantID, matchers...)
 }
 
-func (q *blocksStoreQuerier) LabelNames(ctx context.Context, _ *storage.LabelHints, matchers ...*labels.Matcher) ([]string, annotations.Annotations, error) {
+func (q *blocksStoreQuerier) LabelNames(ctx context.Context, hints *storage.LabelHints, matchers ...*labels.Matcher) ([]string, annotations.Annotations, error) {
 	spanLog, ctx := spanlogger.NewWithLogger(ctx, q.logger, "blocksStoreQuerier.LabelNames")
 	defer spanLog.Span.Finish()
 
@@ -359,7 +359,7 @@ func (q *blocksStoreQuerier) LabelNames(ctx context.Context, _ *storage.LabelHin
 	)
 
 	queryF := func(clients map[BlocksStoreClient][]ulid.ULID, minT, maxT int64) ([]ulid.ULID, error) {
-		nameSets, warnings, queriedBlocks, err := q.fetchLabelNamesFromStore(ctx, clients, minT, maxT, tenantID, convertedMatchers)
+		nameSets, warnings, queriedBlocks, err := q.fetchLabelNamesFromStore(ctx, clients, minT, maxT, tenantID, hints, convertedMatchers)
 		if err != nil {
 			return nil, err
 		}
@@ -377,7 +377,7 @@ func (q *blocksStoreQuerier) LabelNames(ctx context.Context, _ *storage.LabelHin
 	return util.MergeSlices(resNameSets...), resWarnings, nil
 }
 
-func (q *blocksStoreQuerier) LabelValues(ctx context.Context, name string, _ *storage.LabelHints, matchers ...*labels.Matcher) ([]string, annotations.Annotations, error) {
+func (q *blocksStoreQuerier) LabelValues(ctx context.Context, name string, hints *storage.LabelHints, matchers ...*labels.Matcher) ([]string, annotations.Annotations, error) {
 	spanLog, ctx := spanlogger.NewWithLogger(ctx, q.logger, "blocksStoreQuerier.LabelValues")
 	defer spanLog.Span.Finish()
 
@@ -403,7 +403,7 @@ func (q *blocksStoreQuerier) LabelValues(ctx context.Context, name string, _ *st
 	)
 
 	queryF := func(clients map[BlocksStoreClient][]ulid.ULID, minT, maxT int64) ([]ulid.ULID, error) {
-		valueSets, warnings, queriedBlocks, err := q.fetchLabelValuesFromStore(ctx, name, clients, minT, maxT, tenantID, matchers...)
+		valueSets, warnings, queriedBlocks, err := q.fetchLabelValuesFromStore(ctx, name, clients, minT, maxT, tenantID, hints, matchers...)
 		if err != nil {
 			return nil, err
 		}
@@ -999,6 +999,7 @@ func (q *blocksStoreQuerier) fetchLabelNamesFromStore(
 	minT int64,
 	maxT int64,
 	tenantID string,
+	hints *storage.LabelHints,
 	matchers []storepb.LabelMatcher,
 ) ([][]string, annotations.Annotations, []ulid.ULID, error) {
 	var (
@@ -1014,7 +1015,7 @@ func (q *blocksStoreQuerier) fetchLabelNamesFromStore(
 	// Concurrently fetch series from all clients.
 	for c, blockIDs := range clients {
 		g.Go(func() error {
-			req, err := createLabelNamesRequest(minT, maxT, blockIDs, matchers)
+			req, err := createLabelNamesRequest(minT, maxT, blockIDs, hints, matchers)
 			if err != nil {
 				return errors.Wrapf(err, "failed to create label names request")
 			}
@@ -1078,6 +1079,7 @@ func (q *blocksStoreQuerier) fetchLabelValuesFromStore(
 	minT int64,
 	maxT int64,
 	tenantID string,
+	hints *storage.LabelHints,
 	matchers ...*labels.Matcher,
 ) ([][]string, annotations.Annotations, []ulid.ULID, error) {
 	var (
@@ -1093,7 +1095,7 @@ func (q *blocksStoreQuerier) fetchLabelValuesFromStore(
 	// Concurrently fetch series from all clients.
 	for c, blockIDs := range clients {
 		g.Go(func() error {
-			req, err := createLabelValuesRequest(minT, maxT, name, blockIDs, matchers...)
+			req, err := createLabelValuesRequest(minT, maxT, name, blockIDs, hints, matchers...)
 			if err != nil {
 				return errors.Wrapf(err, "failed to create label values request")
 			}
@@ -1183,15 +1185,20 @@ func createSeriesRequest(minT, maxT int64, matchers []storepb.LabelMatcher, skip
 	}, nil
 }
 
-func createLabelNamesRequest(minT, maxT int64, blockIDs []ulid.ULID, matchers []storepb.LabelMatcher) (*storepb.LabelNamesRequest, error) {
+func createLabelNamesRequest(minT, maxT int64, blockIDs []ulid.ULID, hints *storage.LabelHints, matchers []storepb.LabelMatcher) (*storepb.LabelNamesRequest, error) {
+	var limit int64
+	if hints != nil {
+		limit = int64(hints.Limit)
+	}
 	req := &storepb.LabelNamesRequest{
 		Start:    minT,
 		End:      maxT,
 		Matchers: matchers,
+		Limit:    limit,
 	}
 
 	// Selectively query only specific blocks.
-	hints := &hintspb.LabelNamesRequestHints{
+	requestHints := &hintspb.LabelNamesRequestHints{
 		BlockMatchers: []storepb.LabelMatcher{
 			{
 				Type:  storepb.LabelMatcher_RE,
@@ -1201,26 +1208,32 @@ func createLabelNamesRequest(minT, maxT int64, blockIDs []ulid.ULID, matchers []
 		},
 	}
 
-	anyHints, err := types.MarshalAny(hints)
+	anyRequestHints, err := types.MarshalAny(requestHints)
 	if err != nil {
 		return nil, errors.Wrapf(err, "failed to marshal label names request hints")
 	}
 
-	req.Hints = anyHints
+	req.Hints = anyRequestHints
 
 	return req, nil
 }
 
-func createLabelValuesRequest(minT, maxT int64, label string, blockIDs []ulid.ULID, matchers ...*labels.Matcher) (*storepb.LabelValuesRequest, error) {
+func createLabelValuesRequest(minT, maxT int64, label string, blockIDs []ulid.ULID, hints *storage.LabelHints, matchers ...*labels.Matcher) (*storepb.LabelValuesRequest, error) {
+	var limit int64
+	if hints != nil {
+		limit = int64(hints.Limit)
+	}
+
 	req := &storepb.LabelValuesRequest{
 		Start:    minT,
 		End:      maxT,
 		Label:    label,
 		Matchers: convertMatchersToLabelMatcher(matchers),
+		Limit:    limit,
 	}
 
 	// Selectively query only specific blocks.
-	hints := &hintspb.LabelValuesRequestHints{
+	requestHints := &hintspb.LabelValuesRequestHints{
 		BlockMatchers: []storepb.LabelMatcher{
 			{
 				Type:  storepb.LabelMatcher_RE,
@@ -1230,12 +1243,12 @@ func createLabelValuesRequest(minT, maxT int64, label string, blockIDs []ulid.UL
 		},
 	}
 
-	anyHints, err := types.MarshalAny(hints)
+	anyRequestHints, err := types.MarshalAny(requestHints)
 	if err != nil {
 		return nil, errors.Wrapf(err, "failed to marshal label values request hints")
 	}
 
-	req.Hints = anyHints
+	req.Hints = anyRequestHints
 
 	return req, nil
 }